Object Pooling and Money Mechanism Added
Implemented an object pooling system that dynamically populates the pool based on the allowed LuggageTypes defined in the current level's data from the LevelManager. Only prefabs matching the allowed types are added to the pool. A random object from the pool can now be spawned. After, it is returned to the pool using the ReturnObjectToPool method.

Money system has been added. The player's current money is saved and loaded using PlayerPrefs and updates in real-time with earn/spend operations. The money value persists between sessions and can be accessed or reset globally.